[ir] [refactor] Slim ir.h and reduce build time by ~4% (#761)
